More work

+++ src/assets/shaders/scene3d/compileshaders.sh
@@ -16,6 +16,9 @@ USEZIP=0
 # FPC'S RTL-side Windows resource file support and API emulation)   
 USEBIN2C=0 
 
+# Delete the temporary files after compilation
+DELETEAFTERCOMPILE=0
+
 #############################################
 #            Initialization code            #
 #############################################
@@ -1188,6 +1191,10 @@ addMeshFragmentMaterialSourceVariants "mesh" ""
 
 cp -f "${originalDirectory}/bluenoise_1024x1024_rgba8.raw" "${tempPath}/bluenoise_1024x1024_rgba8.raw" || exit 1
 
+cp -f "${originalDirectory}/rain_512.raw" "${tempPath}/rain_512.raw" || exit 1
+
+cp -f "${originalDirectory}/rain_normal_512.raw" "${tempPath}/rain_normal_512.raw" || exit 1
+
 #############################################
 #   Deduplication code for shader binaries  # 
 #############################################
@@ -1357,7 +1364,7 @@ cp packscene3dshaders "${tempPath}/packscene3dshaders"
 cd "${tempPath}"
 
 # Get a sorted list of .spv files, bluenoise_1024x1024_rgba8.raw and virtualsymlinks.json without their full paths
-toCompressFiles=( $((ls *.spv; echo virtualsymlinks.json; echo bluenoise_1024x1024_rgba8.raw) | sort) )
+toCompressFiles=( $((ls *.spv; echo virtualsymlinks.json; echo bluenoise_1024x1024_rgba8.raw; echo rain_512.raw; echo rain_normal_512.raw) | sort) ) # find "${tempPath}" -maxdepth 1 -type f -name "*.c" -printf "%f\n"
 
 if [ $USEZIP -eq 1 ]; then
 
@@ -1556,7 +1563,12 @@ else
 fi
  
 # Delete the temporary directory
-rm -rf "${tempPath}" # what actually deletes also the files in it
+if [ $DELETEAFTERCOMPILE -eq 1 ]; then
+  echo "Deleting temporary directory ${tempPath} . . ."
+  rm -rf "${tempPath}" # what actually deletes also thefiles in it
+else
+  echo "Temporary directory ${tempPath} is not deleted, because DELETEAFTERCOMPILE is set to 0."
+fi
 
 # Done!
